Further unify the autodiff passes.
