+++ source/slang/slang-ir-autodiff-fwd.h
@@ -1,116 +1,19 @@
 // slang-ir-autodiff-fwd.h
 #pragma once
 
-#include "slang-ir.h"
-#include "slang-ir-insts.h"
-#include "slang-compiler.h"
+#include "slang-ir-autodiff-transcriber-base.h"
 
 namespace Slang
 {
 
-    template<typename P, typename D>
-    struct DiffInstPair
-    {
-        P primal;
-        D differential;
-        DiffInstPair() = default;
-        DiffInstPair(P primal, D differential) : primal(primal), differential(differential)
-        {}
-        HashCode getHashCode() const
-        {
-            Hasher hasher;
-            hasher << primal << differential;
-            return hasher.getResult();
-        }
-        bool operator ==(const DiffInstPair& other) const
-        {
-            return primal == other.primal && differential == other.differential;
-        }
-    };
-    
-    typedef DiffInstPair<IRInst*, IRInst*> InstPair;
-
-    
-struct ForwardDerivativeTranscriber
+struct ForwardDerivativeTranscriber : AutoDiffTranscriberBase
 {
-
-    // Stores the mapping of arbitrary 'R-value' instructions to instructions that represent
-    // their differential values.
-    Dictionary<IRInst*, IRInst*>            instMapD;
-
-    // Set of insts currently being transcribed. Used to avoid infinite loops.
-    HashSet<IRInst*>                        instsInProgress;
-
-    // Cloning environment to hold mapping from old to new copies for the primal
-    // instructions.
-    IRCloneEnv                              cloneEnv;
-
-    // Diagnostic sink for error messages.
-    DiagnosticSink*                         sink;
-
-    // Type conformance information.
-    AutoDiffSharedContext*                  autoDiffSharedContext;
-
-    // Builder to help with creating and accessing the 'DifferentiablePair<T>' struct
-    DifferentialPairTypeBuilder*            pairBuilder;
-
-    DifferentiableTypeConformanceContext    differentiableTypeConformanceContext;
-
-    List<InstPair>                          followUpFunctionsToTranscribe;
-
-    SharedIRBuilder* sharedBuilder;
-    // Witness table that `DifferentialBottom:IDifferential`.
-    IRWitnessTable* differentialBottomWitness = nullptr;
-    Dictionary<InstPair, IRInst*> differentialPairTypes;
-
-    ForwardDerivativeTranscriber(AutoDiffSharedContext* shared, SharedIRBuilder* inSharedBuilder)
-        : differentiableTypeConformanceContext(shared), sharedBuilder(inSharedBuilder)
+    ForwardDerivativeTranscriber(AutoDiffSharedContext* shared, SharedIRBuilder* inSharedBuilder, DiagnosticSink* inSink)
+        : AutoDiffTranscriberBase(shared, inSharedBuilder, inSink)
     {
-
     }
 
-    DiagnosticSink* getSink();
-
-    void mapDifferentialInst(IRInst* origInst, IRInst* diffInst);
-
-    void mapPrimalInst(IRInst* origInst, IRInst* primalInst);
-
-    IRInst* lookupDiffInst(IRInst* origInst);
-
-    IRInst* lookupDiffInst(IRInst* origInst, IRInst* defaultInst);
-
-    bool hasDifferentialInst(IRInst* origInst);
-
-    bool shouldUseOriginalAsPrimal(IRInst* origInst);
-
-    IRInst* lookupPrimalInst(IRInst* origInst);
-
-    IRInst* lookupPrimalInst(IRInst* origInst, IRInst* defaultInst);
-
-    bool hasPrimalInst(IRInst* origInst);
-
-    IRInst* findOrTranscribeDiffInst(IRBuilder* builder, IRInst* origInst);
-
-    IRInst* findOrTranscribePrimalInst(IRBuilder* builder, IRInst* origInst);
-
-    IRFuncType* differentiateFunctionType(IRBuilder* builder, IRFuncType* funcType);
-
-    // Get or construct `:IDifferentiable` conformance for a DifferentiablePair.
-    IRWitnessTable* getDifferentialPairWitness(IRInst* inDiffPairType);
-
-    IRType* getOrCreateDiffPairType(IRInst* primalType, IRInst* witness);
-
-    IRType* getOrCreateDiffPairType(IRInst* primalType);
-
-    IRType* differentiateType(IRBuilder* builder, IRType* origType);
-
-    IRType* _differentiateTypeImpl(IRBuilder* builder, IRType* origType);
-
-    IRType* differentiateExtractExistentialType(IRBuilder* builder, IRExtractExistentialType* origType, IRInst*& witnessTable);
-
-    IRType* tryGetDiffPairType(IRBuilder* builder, IRType* primalType);
-
-    InstPair transcribeParam(IRBuilder* builder, IRParam* origParam);
+    virtual IRFuncType* differentiateFunctionType(IRBuilder* builder, IRFuncType* funcType) override;
 
     // Returns "d<var-name>" to use as a name hint for variables and parameters.
     // If no primal name is available, returns a blank string.
@@ -131,8 +34,6 @@ struct ForwardDerivativeTranscriber
     InstPair transcribeLoad(IRBuilder* builder, IRLoad* origLoad);
 
     InstPair transcribeStore(IRBuilder* builder, IRStore* origStore);
-
-    InstPair transcribeReturn(IRBuilder* builder, IRReturn* origReturn);
 
     // Since int/float literals are sometimes nested inside an IRConstructor
     // instruction, we check to make sure that the nested instr is a constant
@@ -158,17 +59,6 @@ struct ForwardDerivativeTranscriber
 
     InstPair transcribeSpecialize(IRBuilder* builder, IRSpecialize* origSpecialize);
 
-    InstPair transcribeLookupInterfaceMethod(IRBuilder* builder, IRLookupWitnessMethod* lookupInst);
-
-    // In differential computation, the 'default' differential value is always zero.
-    // This is a consequence of differential computing being inherently linear. As a 
-    // result, it's useful to have a method to generate zero literals of any (arithmetic) type.
-    // The current implementation requires that types are defined linearly.
-    // 
-    IRInst* getDifferentialZeroOfType(IRBuilder* builder, IRType* primalType);
-
-    InstPair transcribeBlock(IRBuilder* builder, IRBlock* origBlock);
-
     InstPair transcribeFieldExtract(IRBuilder* builder, IRInst* originalInst);
 
     InstPair transcribeGetElement(IRBuilder* builder, IRInst* origGetElementPtr);
@@ -179,9 +69,11 @@ struct ForwardDerivativeTranscriber
 
     InstPair transcribeMakeDifferentialPair(IRBuilder* builder, IRMakeDifferentialPair* origInst);
 
-    InstPair trascribeNonDiffInst(IRBuilder* builder, IRInst* origInst);
-
     InstPair transcribeDifferentialPairGetElement(IRBuilder* builder, IRInst* origInst);
+
+    InstPair transcribeSingleOperandInst(IRBuilder* builder, IRInst* origInst);
+
+    InstPair transcribeWrapExistential(IRBuilder* builder, IRInst* origInst);
 
     // Create an empty func to represent the transcribed func of `origFunc`.
     InstPair transcribeFuncHeader(IRBuilder* inBuilder, IRFunc* origFunc);
@@ -192,19 +84,13 @@ struct ForwardDerivativeTranscriber
     // Transcribe a generic definition
     InstPair transcribeGeneric(IRBuilder* inBuilder, IRGeneric* origGeneric);
 
-    IRInst* transcribe(IRBuilder* builder, IRInst* origInst);
+    virtual InstPair transcribeInstImpl(IRBuilder* builder, IRInst* origInst) override;
 
-    InstPair transcribeInst(IRBuilder* builder, IRInst* origInst);
-};
-
-    struct ForwardDerivativePassOptions
+    virtual IROp getDifferentiableMethodDictionaryItemOp() override
     {
-        // Nothing for now..
-    };
+        return kIROp_ForwardDifferentiableMethodRequirementDictionaryItem;
+    }
 
-    bool processForwardDerivativeCalls(
-        AutoDiffSharedContext*                  autodiffContext,
-        DiagnosticSink*                         sink,
-        ForwardDerivativePassOptions const&     options = ForwardDerivativePassOptions());
+};
 
 }
